Move DQ client and status models to data-quality service
