+++ src/dc43/integration/spark/violation_strategy.py
@@ -12,7 +12,7 @@ except Exception:  # pragma: no cover
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.services.data_quality.backend.engine import ValidationResult
+from dc43.services.data_quality.models import ValidationResult
 
 
 def _merge_pipeline_context(
