+++ src/dc43/services/data_quality/backend/__init__.py
@@ -1,11 +1,11 @@
 """Backend contracts and stubs for the data-quality service."""
 
-from .engine import ExpectationSpec, ValidationResult, evaluate_contract, expectation_specs
+from .engine import ExpectationSpec, evaluate_contract, expectation_specs
 from .interface import DataQualityServiceBackend
 from .local import LocalDataQualityServiceBackend
 from .manager import DataQualityManager
 from dc43.integration.spark.validation import apply_contract
-from ..models import ObservationPayload
+from ..models import ObservationPayload, ValidationResult
 
 __all__ = [
     "DataQualityServiceBackend",
