+++ src/dc43/services/data_quality/backend/interface.py
@@ -6,8 +6,7 @@ from typing import Protocol
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.services.data_quality.backend.engine import ValidationResult
-from dc43.services.data_quality.models import ObservationPayload
+from dc43.services.data_quality.models import ObservationPayload, ValidationResult
 
 
 class DataQualityServiceBackend(Protocol):
