+++ src/dc43/services/data_quality/client/__init__.py
@@ -1,6 +1,10 @@
 """Client interfaces and local implementations for data-quality services."""
 
-from .interface import DataQualityServiceClient
+from .interface import DQClient, DataQualityServiceClient
 from .local import LocalDataQualityServiceClient
 
-__all__ = ["DataQualityServiceClient", "LocalDataQualityServiceClient"]
+__all__ = [
+    "DataQualityServiceClient",
+    "LocalDataQualityServiceClient",
+    "DQClient",
+]
