+++ src/dc43/services/data_quality/engine.py
@@ -1,0 +1,15 @@
+"""Public helpers for evaluating data-quality contracts."""
+
+from dc43.services.data_quality.backend.engine import (
+    ExpectationSpec,
+    apply_contract,
+    evaluate_contract,
+    expectation_specs,
+)
+
+__all__ = [
+    "ExpectationSpec",
+    "apply_contract",
+    "evaluate_contract",
+    "expectation_specs",
+]
