+++ src/dc43/services/governance/backend/__init__.py
@@ -1,6 +1,8 @@
 """Backend contracts and stubs for governance services."""
 
-from .dq import DQClient, DQStatus
+from dc43.services.data_quality.client.interface import DQClient
+from dc43.services.data_quality.models import DQStatus
+
 from .interface import GovernanceServiceBackend
 from .local import LocalGovernanceServiceBackend
