+++ src/dc43/services/governance/backend/interface.py
@@ -6,9 +6,8 @@ from typing import Callable, Mapping, Optional, Protocol, Sequence
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.services.data_quality.backend.engine import ValidationResult
-from dc43.services.governance.backend.dq import DQStatus
-from dc43.services.data_quality.models import ObservationPayload
+from dc43.services.data_quality.models import ObservationPayload, ValidationResult
+from dc43.services.governance.models import DQStatus
 
 from ..models import (
     GovernanceCredentials,
