+++ tests/test_integration.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from typing import Optional
+from typing import Optional, Tuple
 
 import pytest
 
@@ -13,6 +13,7 @@ from open_data_contract_standard.model import (
 )
 
 from dc43.services.contracts.backend.stores import FSContractStore
+from dc43.services.contracts.client import LocalContractServiceClient
 from dc43.integration.spark.io import (
     read_with_contract,
     write_with_contract,
@@ -60,16 +61,18 @@ def make_contract(base_path: str, fmt: str = "parquet") -> OpenDataContractStand
     )
 
 
-def persist_contract(tmp_path: Path, contract: OpenDataContractStandard) -> FSContractStore:
+def persist_contract(
+    tmp_path: Path, contract: OpenDataContractStandard
+) -> Tuple[FSContractStore, LocalContractServiceClient]:
     store = FSContractStore(str(tmp_path / "contracts"))
     store.put(contract)
-    return store
+    return store, LocalContractServiceClient(store)
 
 
 def test_dq_integration_blocks(spark, tmp_path: Path) -> None:
     data_dir = tmp_path / "parquet"
     contract = make_contract(str(data_dir))
-    store = persist_contract(tmp_path, contract)
+    store, contract_service = persist_contract(tmp_path, contract)
     # Prepare data with one enum violation for currency
     data = [
         (1, 101, datetime(2024, 1, 1, 10, 0, 0), 10.0, "EUR"),
@@ -83,7 +86,7 @@ def test_dq_integration_blocks(spark, tmp_path: Path) -> None:
     _, status = read_with_contract(
         spark,
         contract_id=contract.id,
-        contract_store=store,
+        contract_service=contract_service,
         expected_contract_version=f"=={contract.version}",
         enforce=False,
         governance_service=governance,
@@ -100,7 +103,7 @@ def test_dq_integration_blocks(spark, tmp_path: Path) -> None:
 def test_write_violation_blocks_by_default(spark, tmp_path: Path) -> None:
     dest_dir = tmp_path / "dq"
     contract = make_contract(str(dest_dir))
-    store = persist_contract(tmp_path, contract)
+    store, contract_service = persist_contract(tmp_path, contract)
     df = spark.createDataFrame(
         [
             (1, 101, datetime(2024, 1, 1, 10, 0, 0), 10.0, "EUR"),
@@ -112,7 +115,7 @@ def test_write_violation_blocks_by_default(spark, tmp_path: Path) -> None:
     result, status = write_with_contract(
         df=df,
         contract_id=contract.id,
-        contract_store=store,
+        contract_service=contract_service,
         expected_contract_version=f"=={contract.version}",
         mode="overwrite",
         enforce=False,
@@ -131,14 +134,14 @@ def test_write_violation_blocks_by_default(spark, tmp_path: Path) -> None:
 def test_write_validation_result_on_mismatch(spark, tmp_path: Path):
     dest_dir = tmp_path / "out"
     contract = make_contract(str(dest_dir))
-    store = persist_contract(tmp_path, contract)
+    store, contract_service = persist_contract(tmp_path, contract)
     # Missing required column 'currency' to trigger validation error
     data = [(1, 101, datetime(2024, 1, 1, 10, 0, 0), 10.0)]
     df = spark.createDataFrame(data, ["order_id", "customer_id", "order_ts", "amount"])
     result = write_with_contract(
         df=df,
         contract_id=contract.id,
-        contract_store=store,
+        contract_service=contract_service,
         expected_contract_version=f"=={contract.version}",
         mode="overwrite",
         enforce=False,  # continue writing despite mismatch
@@ -167,7 +170,7 @@ def test_write_warn_on_path_mismatch(spark, tmp_path: Path):
     expected_dir = tmp_path / "expected"
     actual_dir = tmp_path / "actual"
     contract = make_contract(str(expected_dir))
-    store = persist_contract(tmp_path, contract)
+    store, contract_service = persist_contract(tmp_path, contract)
     data = [
         (1, 101, datetime(2024, 1, 1, 10, 0, 0), 10.0, "EUR"),
     ]
@@ -178,7 +181,7 @@ def test_write_warn_on_path_mismatch(spark, tmp_path: Path):
     result = write_with_contract(
         df=df,
         contract_id=contract.id,
-        contract_store=store,
+        contract_service=contract_service,
         expected_contract_version=f"=={contract.version}",
         path=str(actual_dir),
         mode="overwrite",
@@ -191,7 +194,7 @@ def test_write_path_version_under_contract_root(spark, tmp_path: Path, caplog):
     base_dir = tmp_path / "data"
     contract_path = base_dir / "orders_enriched.parquet"
     contract = make_contract(str(contract_path))
-    store = persist_contract(tmp_path, contract)
+    store, contract_service = persist_contract(tmp_path, contract)
     data = [
         (1, 101, datetime(2024, 1, 1, 10, 0, 0), 10.0, "EUR"),
     ]
@@ -204,7 +207,7 @@ def test_write_path_version_under_contract_root(spark, tmp_path: Path, caplog):
         result = write_with_contract(
             df=df,
             contract_id=contract.id,
-            contract_store=store,
+            contract_service=contract_service,
             expected_contract_version=f"=={contract.version}",
             path=str(target),
             mode="overwrite",
@@ -217,7 +220,7 @@ def test_write_path_version_under_contract_root(spark, tmp_path: Path, caplog):
 def test_read_warn_on_format_mismatch(spark, tmp_path: Path, caplog):
     data_dir = tmp_path / "json"
     contract = make_contract(str(data_dir), fmt="parquet")
-    store = persist_contract(tmp_path, contract)
+    store, contract_service = persist_contract(tmp_path, contract)
     data = [
         (1, 101, datetime(2024, 1, 1, 10, 0, 0), 10.0, "EUR"),
     ]
@@ -230,7 +233,7 @@ def test_read_warn_on_format_mismatch(spark, tmp_path: Path, caplog):
         read_with_contract(
             spark,
             contract_id=contract.id,
-            contract_store=store,
+            contract_service=contract_service,
             expected_contract_version=f"=={contract.version}",
             format="json",
             enforce=False,
@@ -244,7 +247,7 @@ def test_read_warn_on_format_mismatch(spark, tmp_path: Path, caplog):
 def test_write_warn_on_format_mismatch(spark, tmp_path: Path, caplog):
     dest_dir = tmp_path / "out"
     contract = make_contract(str(dest_dir), fmt="parquet")
-    store = persist_contract(tmp_path, contract)
+    store, contract_service = persist_contract(tmp_path, contract)
     data = [
         (1, 101, datetime(2024, 1, 1, 10, 0, 0), 10.0, "EUR"),
     ]
@@ -256,7 +259,7 @@ def test_write_warn_on_format_mismatch(spark, tmp_path: Path, caplog):
         result = write_with_contract(
             df=df,
             contract_id=contract.id,
-            contract_store=store,
+            contract_service=contract_service,
             expected_contract_version=f"=={contract.version}",
             path=str(dest_dir),
             format="json",
@@ -276,7 +279,7 @@ def test_write_warn_on_format_mismatch(spark, tmp_path: Path, caplog):
 def test_write_split_strategy_creates_auxiliary_datasets(spark, tmp_path: Path):
     base_dir = tmp_path / "split"
     contract = make_contract(str(base_dir))
-    store = persist_contract(tmp_path, contract)
+    store, contract_service = persist_contract(tmp_path, contract)
     data = [
         (1, 101, datetime(2024, 1, 1, 10, 0, 0), 10.0, "EUR"),
         (2, 102, datetime(2024, 1, 2, 10, 0, 0), 15.5, "INR"),
@@ -290,7 +293,7 @@ def test_write_split_strategy_creates_auxiliary_datasets(spark, tmp_path: Path):
     result = write_with_contract(
         df=df,
         contract_id=contract.id,
-        contract_store=store,
+        contract_service=contract_service,
         expected_contract_version=f"=={contract.version}",
         mode="overwrite",
         enforce=False,
@@ -319,7 +322,7 @@ def test_write_dq_violation_reports_status(spark, tmp_path: Path):
     contract = make_contract(str(dest_dir))
     # Tighten quality rule to trigger a violation for the sample data below.
     contract.schema_[0].properties[3].quality = [DataQuality(mustBeGreaterThan=100)]
-    store = persist_contract(tmp_path, contract)
+    store, contract_service = persist_contract(tmp_path, contract)
 
     data = [
         (1, 101, datetime(2024, 1, 1, 10, 0, 0), 50.0, "EUR"),
@@ -335,7 +338,7 @@ def test_write_dq_violation_reports_status(spark, tmp_path: Path):
     result, status = write_with_contract(
         df=df,
         contract_id=contract.id,
-        contract_store=store,
+        contract_service=contract_service,
         expected_contract_version=f"=={contract.version}",
         mode="overwrite",
         enforce=False,
@@ -352,7 +355,7 @@ def test_write_dq_violation_reports_status(spark, tmp_path: Path):
         write_with_contract(
             df=df,
             contract_id=contract.id,
-            contract_store=store,
+            contract_service=contract_service,
             expected_contract_version=f"=={contract.version}",
             mode="overwrite",
             enforce=True,
@@ -375,6 +378,7 @@ def test_write_keeps_existing_link_for_contract_upgrade(spark, tmp_path: Path):
 
     store = FSContractStore(str(tmp_path / "upgrade_contracts"))
     store.put(contract_v1)
+    contract_service = LocalContractServiceClient(store)
     governance = build_local_governance_service(store)
     upgrade_locator = StaticDatasetLocator(
         dataset_version="2024-01-01",
@@ -383,7 +387,7 @@ def test_write_keeps_existing_link_for_contract_upgrade(spark, tmp_path: Path):
     _, status_ok = write_with_contract(
         df=df_ok,
         contract_id=contract_v1.id,
-        contract_store=store,
+        contract_service=contract_service,
         expected_contract_version=f"=={contract_v1.version}",
         mode="overwrite",
         enforce=False,
@@ -412,7 +416,7 @@ def test_write_keeps_existing_link_for_contract_upgrade(spark, tmp_path: Path):
 def test_governance_service_persists_draft_context(spark, tmp_path: Path) -> None:
     dest_dir = tmp_path / "handles"
     contract = make_contract(str(dest_dir))
-    store = persist_contract(tmp_path, contract)
+    store, contract_service = persist_contract(tmp_path, contract)
 
     # Missing the 'currency' column to trigger a draft proposal.
     data = [
@@ -430,7 +434,7 @@ def test_governance_service_persists_draft_context(spark, tmp_path: Path) -> Non
     result = write_with_contract(
         df=df,
         contract_id=contract.id,
-        contract_store=store,
+        contract_service=contract_service,
         expected_contract_version=f"=={contract.version}",
         mode="overwrite",
         enforce=False,
